+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Hot reloading now only reloads the survey, not the whole page.
 * Added `UrlParameters` argument to the `Survey` class to specify what parameters should be saved in the database.
+* Added `pauseBuild` arg to `SurveyModel.build()` to prevent the survey from being built when created.
+* Added `surveyFromJson()` function to create a survey from a json file made with the visual creator.
 
 ## 0.3.2